:truck: rename project structure
