+++ src/main/kotlin/com/highv/ecommerce/domain/admin/dto/BlackListResponse.kt
@@ -1,0 +1,8 @@
+package com.highv.ecommerce.domain.admin.dto
+
+data class BlackListResponse(
+    val nickname: String,
+    val email: String,
+    val sanctionsCount: Int,
+    val isSanctioned: Boolean
+)
