+++ src/main/kotlin/com/highv/ecommerce/domain/admin/dto/CreateBlackListRequest.kt
@@ -1,0 +1,6 @@
+package com.highv.ecommerce.domain.admin.dto
+
+data class CreateBlackListRequest(
+    val sellerId: Long,
+    val email: String
+)
