+++ src/main/kotlin/com/highv/ecommerce/domain/admin/repository/BlackListRepository.kt
@@ -1,0 +1,9 @@
+package com.highv.ecommerce.domain.admin.repository
+
+import com.highv.ecommerce.domain.admin.entity.BlackList
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface BlackListRepository : JpaRepository<BlackList, Long> {
+    fun existsByEmail(email: String): Boolean
+    fun findByEmail(email: String): BlackList?
+}
