--- src/main/kotlin/com/highv/ecommerce/domain/backoffice/admin/dto/BlackListResponse.kt
@@ -1,8 +1,0 @@
-package com.highv.ecommerce.domain.backoffice.admin.dto
-
-data class BlackListResponse(
-    val nickname: String,
-    val email: String,
-    val sanctionsCount: Int,
-    val isSanctioned: Boolean
-)
