--- src/main/kotlin/com/highv/ecommerce/domain/backoffice/admin/dto/CreateBlackListRequest.kt
@@ -1,6 +1,0 @@
-package com.highv.ecommerce.domain.backoffice.admin.dto
-
-data class CreateBlackListRequest(
-    val sellerId: Long,
-    val email: String
-)
