--- src/main/kotlin/com/highv/ecommerce/domain/product/dto/CreateReviewRequest.kt
@@ -1,6 +1,0 @@
-package com.highv.ecommerce.domain.product.dto
-
-data class CreateReviewRequest(
-    val rate: Float,
-    val content: String
-)
