--- src/main/kotlin/com/highv/ecommerce/domain/product/dto/UpdateReviewRequest.kt
@@ -1,6 +1,0 @@
-package com.highv.ecommerce.domain.product.dto
-
-data class UpdateReviewRequest(
-    val rate : Float,
-    val content : String
-)
