+++ src/main/kotlin/com/highv/ecommerce/domain/product/entity/Product.kt
@@ -1,7 +1,7 @@
 package com.highv.ecommerce.domain.product.entity
 
 import com.highv.ecommerce.domain.backoffice.entity.ProductBackOffice
-import com.highv.ecommerce.domain.shop.entity.Shop
+import com.highv.ecommerce.domain.seller.shop.entity.Shop
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
