+++ src/main/kotlin/com/highv/ecommerce/domain/review/dto/CreateReviewRequest.kt
@@ -1,0 +1,6 @@
+package com.highv.ecommerce.domain.review.dto
+
+data class CreateReviewRequest(
+    val rate: Float,
+    val content: String
+)
