+++ src/main/kotlin/com/highv/ecommerce/domain/review/dto/ReviewResponse.kt
@@ -1,0 +1,8 @@
+package com.highv.ecommerce.domain.review.dto
+
+data class ReviewResponse(
+    val id: Long,
+    val rate: Float,
+    val content: String,
+
+)
