+++ src/main/kotlin/com/highv/ecommerce/domain/review/dto/UpdateReviewRequest.kt
@@ -1,0 +1,6 @@
+package com.highv.ecommerce.domain.review.dto
+
+data class UpdateReviewRequest(
+    val rate : Float,
+    val content : String
+)
