+++ src/main/kotlin/com/highv/ecommerce/domain/review/repository/ReviewRepository.kt
@@ -1,0 +1,12 @@
+package com.highv.ecommerce.domain.review.repository
+
+import com.highv.ecommerce.domain.review.entity.Review
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface ReviewRepository: JpaRepository<Review, Long> {
+    fun findAllByProductId(productId: Long): List<Review>
+    fun findByIdAndBuyerId(Id: Long, buyerId: Long): Review? //수정필요
+    fun findAllByBuyerId(buyerId: Long): List<Review>
+}
