+++ src/main/kotlin/com/highv/ecommerce/domain/seller/shop/dto/CreateShopRequest.kt
@@ -1,0 +1,7 @@
+package com.highv.ecommerce.domain.seller.shop.dto
+
+data class CreateShopRequest(
+    val name: String,
+    val description: String,
+    val shopImage: String
+)
