--- src/main/kotlin/com/highv/ecommerce/domain/shop/dto/CreateShopRequest.kt
@@ -1,7 +1,0 @@
-package com.highv.ecommerce.domain.shop.dto
-
-data class CreateShopRequest(
-    val name: String,
-    val description: String,
-    val shopImage: String
-)
