--- src/main/kotlin/com/highv/ecommerce/domain/shop/entity/Shop.kt
@@ -1,29 +1,0 @@
-package com.highv.ecommerce.domain.shop.entity
-
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.GenerationType
-import jakarta.persistence.Id
-
-@Entity
-class Shop(
-    @Column(name = "sellerId")
-    var sellerId: Long,
-
-    @Column(name = "name")
-    var name: String,
-
-    @Column(name = "description")
-    var description: String,
-
-    @Column(name = "shop_image")
-    var shopImage: String,
-
-    @Column(name = "rate")
-    var rate: Float,
-) {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Long? = null
-}
