+++ src/test/kotlin/com/highv/ecommerce/coupon/service/CouponServiceTest.kt
@@ -14,7 +14,7 @@ import com.highv.ecommerce.domain.coupon.service.CouponService
 import com.highv.ecommerce.domain.item_cart.repository.ItemCartRepository
 import com.highv.ecommerce.domain.product.entity.Product
 import com.highv.ecommerce.domain.product.repository.ProductRepository
-import com.highv.ecommerce.domain.shop.entity.Shop
+import com.highv.ecommerce.domain.seller.shop.entity.Shop
 import com.highv.ecommerce.infra.security.UserPrincipal
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
