+++ src/test/kotlin/com/highv/ecommerce/login/SignUpServiceTest.kt
@@ -8,7 +8,7 @@ import com.highv.ecommerce.domain.seller.dto.CreateSellerRequest
 import com.highv.ecommerce.domain.seller.entity.Seller
 import com.highv.ecommerce.domain.seller.repository.SellerRepository
 import com.highv.ecommerce.domain.seller.service.SellerService
-import com.highv.ecommerce.s3.config.S3Manager
+import com.highv.ecommerce.infra.s3.S3Manager
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import io.mockk.every
