+++ src/test/kotlin/com/highv/ecommerce/order_detail/service/BuyerOrderServiceTest.kt
@@ -13,7 +13,7 @@ import com.highv.ecommerce.domain.order_details.service.OrderDetailsService
 import com.highv.ecommerce.domain.order_master.entity.OrderMaster
 import com.highv.ecommerce.domain.order_master.repository.OrderMasterRepository
 import com.highv.ecommerce.domain.product.entity.Product
-import com.highv.ecommerce.domain.shop.entity.Shop
+import com.highv.ecommerce.domain.seller.shop.entity.Shop
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
 import io.mockk.every
