+++ src/test/kotlin/com/highv/ecommerce/order_master/service/OrderMasterServiceService.kt
@@ -19,7 +19,7 @@ import com.highv.ecommerce.domain.order_master.entity.OrderMaster
 import com.highv.ecommerce.domain.order_master.repository.OrderMasterRepository
 import com.highv.ecommerce.domain.order_master.service.OrderMasterService
 import com.highv.ecommerce.domain.product.entity.Product
-import com.highv.ecommerce.domain.shop.entity.Shop
+import com.highv.ecommerce.domain.seller.shop.entity.Shop
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import io.mockk.every
